Move middleware code into api_routes for faster compilation
