+++ crates/api/src/local_user/logout.rs
@@ -1,7 +1,10 @@
-use crate::read_auth_token;
 use activitypub_federation::config::Data;
 use actix_web::{cookie::Cookie, HttpRequest, HttpResponse};
-use lemmy_api_common::{context::LemmyContext, utils::AUTH_COOKIE_NAME, SuccessResponse};
+use lemmy_api_common::{
+  context::LemmyContext,
+  utils::{read_auth_token, AUTH_COOKIE_NAME},
+  SuccessResponse,
+};
 use lemmy_db_schema::source::login_token::LoginToken;
 use lemmy_db_views::structs::LocalUserView;
 use lemmy_utils::error::{LemmyErrorType, LemmyResult};
