+++ crates/api_common/Cargo.toml
@@ -61,8 +61,8 @@ reqwest = { workspace = true, optional = true }
 ts-rs = { workspace = true, optional = true }
 moka = { workspace = true, optional = true }
 anyhow.workspace = true
-actix-web = { workspace = true, optional = true }
 enum-map = { workspace = true }
+actix-web = { workspace = true, optional = true }
 urlencoding = { workspace = true }
 mime = { version = "0.3.17", optional = true }
 mime_guess = "2.0.5"
@@ -72,6 +72,8 @@ webpage = { version = "2.0", default-features = false, optional = true, features
 ] }
 encoding_rs = { version = "0.8.35", optional = true }
 jsonwebtoken = { version = "9.3.0", optional = true }
+actix-web-httpauth = "0.8.2"
+webmention = "0.6.0"
 
 [dev-dependencies]
 serial_test = { workspace = true }
