+++ crates/api_crud/Cargo.toml
@@ -27,7 +27,6 @@ futures.workspace = true
 uuid = { workspace = true }
 anyhow.workspace = true
 chrono.workspace = true
-webmention = "0.6.0"
 accept-language = "3.1.0"
 regex = { workspace = true }
 serde_json = { workspace = true }
