+++ crates/routes/Cargo.toml
@@ -32,5 +32,16 @@ serde = { workspace = true }
 url = { workspace = true }
 tracing = { workspace = true }
 tokio = { workspace = true }
+futures-util.workspace = true
 http.workspace = true
+diesel.workspace = true
+diesel-async.workspace = true
+clokwerk = "0.4.0"
+prometheus = { version = "0.13.4", features = ["process"] }
 rss = "2.0.11"
+actix-web-prom = "0.9.0"
+actix-cors = "0.7.0"
+
+[dev-dependencies]
+pretty_assertions.workspace = true
+serial_test.workspace = true
