+++ crates/routes/src/lib.rs
@@ -1,17 +1,6 @@
-use lemmy_api_common::{claims::Claims, context::LemmyContext, utils::check_user_valid};
-use lemmy_db_views::structs::LocalUserView;
-use lemmy_utils::error::LemmyResult;
-
 pub mod feeds;
 pub mod images;
+pub mod middleware;
 pub mod nodeinfo;
+pub mod utils;
 pub mod webfinger;
-
-#[tracing::instrument(skip_all)]
-async fn local_user_view_from_jwt(jwt: &str, context: &LemmyContext) -> LemmyResult<LocalUserView> {
-  let local_user_id = Claims::validate(jwt, context).await?;
-  let local_user_view = LocalUserView::read(&mut context.pool(), local_user_id).await?;
-  check_user_valid(&local_user_view.person)?;
-
-  Ok(local_user_view)
-}
